Moves the config directory into the src directory
