+++ routes/api/users.js
@@ -3,7 +3,7 @@ const router = express.Router();
 
 const User = require("../../models/User");
 const bcrypt = require("bcryptjs");
-const keys = require('../../config/keys');
+const keys = require('../../frontend/src/config/keys');
 const jwt = require("jsonwebtoken");
 
 const validateRegisterInput = require("../../validation/register");
